graphics,tools: Change BufferInitializer to have access only to a Buffer object
Change BufferInitializer to have access only to a Buffer object, instead of a
Buffer object and an EGLClientBuffer handle. This gives up some versatility in
the handling of the buffer contents by the BufferInitializer, but ensures that
the BufferInitializer can't use the buffer's EGL resources in an way that is
incompatible with the rest of Mir.