gbm: Use a separate BufferTextureBinder object to handle texture binding functionality
Use a separate BufferTextureBinder object to handle texture binding
functionality, decoupling the GBMBuffer from the actual mechanism that
performs the binding.

In addition to the design benefits, this change fixes a series of test
failures we had when running the tests repeatedly in the same process
(because EGL extension pointer initialization was done only once per
process).