tests/gbm: Fix hanging tests
The threads_switch_loop_master and threads_loop_master_notifies_non_master
tests were created with the assumption that the first thread is the loop
master, but erroneously allowed either thread to become the loop master.
When the second thread happened to become the loop master, the tests hung.
This change forces the first thread to become the loop master.