MirWaitHandle: Add expect_result() method and some unit tests.
expect_result() is required to support assymmetric usage patterns of a
MirWaitHandle like:
    mir_foo(x, A);
    mir_foo(x, B);
    mir_wait_for(mir_foo(x, C));
where mir_foo always reuses the same MirWaitHandle for x.

Previously, mir_wait_for would return as soon as any call to mir_foo finished,
meaning only A or B had completed. But clearly what we want is to wait for C.
That is what expect_result allows as to support as new client functions like
mir_foo are added.