Prototype thread-safety for MirConnection and MirSurface, and particularly their callbacks.
Currently each object (MirSurface or MirConnection) has its own recursive
mutex. This mutex protects the object's members as well as its callbacks
(any callbacks whose first parameter is a pointer to that object).

This allows significant concurrency as independent connections and surfaces
will not block each other. But it also ensures that the callbacks for an
object can only execute one-at-a-time, and never at the same time as the
object is being used in another thread.