Rename compositor::Renderables --> compositor::Scene
What is a "Renderables"? It's the interface by which we know the scene graph,
or the model of how all the surfaces relate to each other. Right, then call it
a scene graph... but the word "graph" might imply a particular structure. And
we don't want to imply such things in an interface. So just call it "scene".
That makes slightly more sense than imagining what a "Renderables" (singular)
is.