client: Remove the timeout for detecting server crashes
There are valid cases when the server blocks for an indeterminate
amount of time during a client request, like when a next_buffer request
blocks because a surface is not visible or the server is paused. In such
cases, the timeout mechanism we used to detect server crashes caused
the clients to fail. Furthermore, the clients seem able to detect server
crashes without the timeout, rendering it needless anyway.