gbm: Don't mess up the VT mode on setup failure
Only restore the previous VT mode during shutdown if it was VT_AUTO. VT_PROCESS
mode is always bound to the calling process, so "restoring" VT_PROCESS will not
work; it will just bind the notification signals to our process again.

Not "restoring" VT_PROCESS also ensures we don't mess up the VT state of the
previous controlling process, in case it had set VT_PROCESS and we fail during
setup. This could happen, for example, when trying to start Mir from inside an
X session.