Replace some twine in cross-compile-chroot.sh
By exporting the PKG_CONFIG_* environment variables we get pkg-config to output the
correct values for cflags and ldflags.

Doing so causes the cross-compiler to pick up the libc libraries
in the chroot, which aren't appropriate (they contain absolute paths),
so don't unpack the libc bits in the partial chroot.