Updates the InputWindowInfo found at InputWindowHandles everytime the InputDispatcher tries to querry the object for hit testing.
So far that information was only updated during creation (because we triggered that in the mir implemantation) and when focus changes occur - that might be the reason why unity8 never experienced these issues. This fix kind of makes updateInfo and the interface of InputWindowInfo even more questionable. Why have those members public? Why update these only under certain circumstances..