Workaround for Greeter turning the screen on.
This is just a workaround for the issue with the split greeter. While display is off another greeter is spawned to be shown as soon as display is turned on again. A greeter is a nested mir server. On initialization nested servers send the initial display configuration to the hosting server. Obviously that configuration contains power_mode == MirPowerMode::mir_power_mode_on.
Usc sets the focus onto the recently started greeter as soon as the session has a surface with buffers in. On that focus change the the stored display configuration (given they have one) is applied. Which turns the display on.

This change disables sending the first/initial nested display configuration.
We clearly need a better solution for that issue, and we should remove that workaround as soon as we have one.