demo-shell: Select which monitor to rotate or change mode on using the mouse cursor location.
FIXME: This seems to reveal pre-existing bugs where after some re-configures
the pointer gets stranded and can't move to one or more monitors. This bug is
the reason why the change has not also been applied to power_mode, because
turning a display off makes it impossible to reach with the cursor and thus
impossible to turn back on.