Removal of the InputRegistrar interface and introduction of input reception mode in mr::input::Surface
With the new scene observer the input registrar interface is no longer necessary to inform the input stack about added/removed surfaces and channels. Since the scene observer only provides a surface pointer to the caller the input reception mode was turned into an attribute of the input::Surface. With that changes to the property are also notified to surface observers.