Automate checking for ABI breakage.
This may well be over-sensitive to header changes that don't constitute
ABI changes, but I think better that than being under-sensitive.

You can find out which header(s) are causing the failre by running:
  sha1sum -c server-ABI-23-sha1sums
  sha1sum -c client-ABI-8-sha1sums

If the offending change is a real ABI break, bump it! Then regardless of
whether it was a break, generate a new checksum for the file using "sha1sum"
and insert it into {client,server}-ABI-?-sha1sums.