Fixes an update problem that causes an exception inside ASIO.
The problem occurs when the response of a client races with another send operation about the mia::InputSender::ActiveTransfer object. Previously a reponse leads to deregistration of the fd. That operation is implemented by first removing the ActiveTransfer from the internal registry before it gets destroyed and unregistered. Meanwhile the sending thread forces the creation of another ActiveTransfer with the same file descriptor. This operation fails at epoll_ctl, since it attempts to add the fd another time.
  
This fix changes InputSender to register an FDHandler when on first send attempt, and unregisters the file descriptor when the surface removed. Additionally the unregistration on destruction is done with the sender_mutex locked. Which avoids a similar scenario when a surface gets removed while one with the same fd gets added and used.