Increase timeout for (yet!) another test
This seems to be semi-regularly taking 2.5sec on CI, and the other
test is very nearly hitting the 2s limit, too. Bump to 60s as this
timeout is only for the failure case and it's much faster to not fail
than to spuriously hit the timeout and require a whole new CI run.