Return false from Dispatchable::dispatch() to indicate no further events.
Once the remote end has closed the connection a fd will remain in state
fd_event::remote_closed indefinitely. Likewise, a file descriptor at EOF
will indefinitely be readable.

At the same time it's unsafe to ::close the relevant fd. For a start,
there's no guarantee that someone hasn't dup()ed it, or kept a reference
to the mir::Fd that will now close some random file descriptor when it
loses scope. Even if neither of those are true, close()ing a file descriptor
when some other thread is in select(), poll(), or epoll_wait() invokes
undefined behaviour.

Let the Dispatchable return false from dispatch to signal that it should
no longer be called, and punt that problem to whatever is dispatching it.