Adds mir::Flags to common headers.
mir::Flags is a utility to allow treating the type safe scoped enums as a set of masks or bit flags that can be combined in a type safe way. mir::Flags also works with normal enumerations. Additionally it provides utilities to print enumeration values as strings. For that purpose a variadic macro is provided.