Fix race in MultiplexingDispatcher::remove_watch()
We tried to optimise the case where there are no threads in dispatch() by
immediately deleting the relevant Dispatchable.

Sadly, this was racy in the presence of multiple calls to remove_watch():
after the first call to remove_watch() we can no longer determine whether
there are any threads in dispatch() using the Dispatchable we want to remove.

Add a test, and remove the optimisation to fix it.