Use a pthread read-write lock in md::MultiDispatchable rather than atomics+GC.
It turns out that pthread_rwlock is just as fast and is significantly
more obviously correct.

The existing mir::RecursiveReadWriteMutex is half as fast and has a
lot of behaviours we don't need here; recursive locking and being
able to take both a read-lock and a write-lock in the same thread.