Include the omitted parts back into the bugfix
Back then when the cursor issue of radeons and kaveris was handled I omitted this check to simplify the fix. This part was necessary to make GBM return correct stride sizes. GBM seems to stick to the sizeprovided on construction, and does not verify what buffer was returned by the driver. Yes this is broken in GBM, but we are currently not alone with this workaround.