Make the input resampling/prediction rate configurable by environment variable MIR_INPUT_RATE=Hz
We expect resampling won't be around for much longer as Qt can now do its
own. Also, clever apps should be allowed to get the raw input stream if
they know how to use it.

But rather than changing the default behaviour yet, just make it easy
to experiment with...