Change EventHub to provide a mir::Fd for Integration in MultiplexingDispatchables
The motivation of the changes is to prepare EventHub & InutReader to become members of a evdev input platform.
InputReaderDispatchable: new Dispatchable implementation that wrapps EventHub and InputReader
Changes to DefaultInputManager: the new input manager introduced with the stub input platform now uses InputReaderDispatchable to also drive the input devices opened by EventHub, and processed by InputReader, that way mir::input::android::InputManager becomes obsolete.
Changes to SimpleDispatchThread: The switch to DefaultInputManager uncovered a few missing features. The ability to avoid the signal deactivation, and handling execeptions thrown from dispatachables. Since SimpleDispatchThread is part of common, and may also be used inside the client, this is solved by passing a functor to enclose the event loop.