Rework thread-safety of ActionQueue.
Instead protecting against data races by throwing an exception (and likely killing
the process), detect when we've lost a race and don't attempt to process the event.

This slightly changes the ActionQueue semantics - actions can now be potentially
processed in parallel, and each call to dispatch() processes a single action, rather
than draining the queue.