Ensure that there's always at least one loadable DSO to test.
If we're building, then we obviously have a compiler and linker for the current architecture.
Use them to generate a libthis-arch.so alongside our libi386.so, libamd64.so etc in
test_data so that there's always a module that will successfully load.

This removes the final arch dependency on our testsuite.

(Apart from the failure to handle big-endian architectures, but that's not
a test-suite problem ☺ )