Resolve entirely-non-obvious and harmless data race.
Why do we have two mir::protobuf::Void objects in the MirConnection object, I hear you ask?
Well, because one of them is used as a sent parameter of the disconnect() call, and so is
accessed outside the RPC read lock.

This means that to appease ThreadSanitizer you need to use void_result for void results
and ignored for void parameters.

The More You Know™