Decouple ~GSourceHandle from callback guarantees.
~GSourceHandle, via (among other things) pre_destruction_hook, provided the guarantee that
the destructor would not complete until there were no threads in the callback and that
no further threads would enter the callback.

This guarantee is problematic for AlarmImpl::reschedule_* because providing those guarantees
makes it difficult to avoid deadlocking while providing the behaviour Alarm requires.

It is also an unnecessary guarantee for AlarmImpl; any attempt to use reschedule_*
is inherently racing against real-time, and the mainloop ensures the callback
isn't called reentrantly.

So make ‘wait until we're definitely not getting a callback’ an explicit call
on GSourceHandle.