remove the test for x11-input
Since stub input is now only set during tests that need to use fake input devices the check for the server platform name is no longer necessary. As soon as an input platform is configured the android input stack is disabled.

Still this is not the final solution - Instead we should probe for available platforms and also allow loading multiple platforms.