Query HWC for display connection if hwc->set fails.
Avoid using hotplug events to ignore commit errors; instead have the wrapper query the external display connection status and issue a DisplayDisconnectedException in such cases. The DisplayGroup can then choose to ignore that specific exception.

This only minimizes the occurences of lp:1498550, as there are still situations in which hwc->set fails due to unplugging the external display but hwc still reports the display as connected. The only recovery in this situation seems to be tearing down HWC.