Make SystemCompositorWindowManager.if_no_surface_posts_client_never_gets_focus resiliant to non-focus events being sent.
If a mock function has an EXPECT_CALL set for it, even if it's .Times(0), then any calls
that *don't* match that specifier become errors.

There's no reason to fail that test if the surface receives a non-focus event, and there's
no reason to add expectations for any other surface events we might choose to send now
or in the future.