Rework Modifier tracking to combine modifiers within
The platform is no longer in charge of the modifier. Instead the new input stack tracks it when the events are received. To reduce the complexity of InputDeviceHub this behavior and similar parts are moved to mir::input::Seat which represents a group of input devices that belong to a user session. (Currently all of them). 

Collateral changes:
 - allow changing the modifier of MirEvent after creation -> additional symbols to 'hidden' C++ client API
 - reduction of EventBuilder interface
 - moving of modifier utils to server