Switch to UniqueModulePtr in DisplayServerConfiguration
With this the lifetime of platform shared module is tied to the lifetime of the graphics::Platform instance used by the server. So we can get rid of the global static SharedLibrary in mirserver.

The platform plays factory for Display, PlatformIPCOperations and GraphicsBufferAllocator. These instances are still created as plain shared ptrs. DisplayServer guarantees that the Platform outlives those objects.