Untangle stanza name from ABI numbering
As of the current doc/dso_versioning_guide the stanza version suffix is changed to contain the 
version the symbol was introduced (or last compatible mir version). To avoid confusing stanza
names with abi numbering this change separates this into two cmake variables, such that our scripting
for abi dumps and update_package_abis does not require changes.