Unify the cursor position within a seat
This move changes the responsibility to track the cursor position out of the platforms to the mir::input::Seat. This reduces the EventBuilder interface and removes the InputSink::confine_cursor method - which was only added since there was no relative axis information in the beginnings of MirEvent.

The fixes cursor jumps when switching between mouse and keyboard.