Work around unpredictable scroll magnitude. (LP: #1522673)
Knowing that all desktop mice implement scroll wheel as discrete
button 4/5 events, a robust client should do this anyway. The existence
of a magnitude is an illusion introduced by libinput, so ignore it.

The direction is still backwards but we'll fix that in one of the
fix branches.

I still wonder why input APIs insist on creating this problem with
a scroll "axis". All wheel mice I've ever seen emitted discrete
button events. It would be clearer for everyone if we just saw button
4/5 events. Just as I have to synthesize in Xmir already...