Don't use memcheck on the ClientLatency acceptance test. Because valgrind is too slow to adequately emulate the multi-process performance we're testing for and will cause failures.
This is a kind of workaround for LP: #1522031 which itself probably
won't and shouldn't be "fixed".

Also don't move the offending test out of acceptance and into performance
tests because it's a crucial test for a lot of work going on of late and we
need it to catch mistakes as we make them.