MirSurface: do not hold mutex while creating buffer stream
During MirSurface::created, the mutex should be aqcuired only to protext writes to the MirSurface::buffer_stream and MirSurface::attrib_cache.
Read access to MirSurface::surface do not need a lock as the writes for it have happened externally by ipc/rpc code and MirSurface::created is invoked within the same thread.