Add input::Seat and move implementation details into separate class
A Seat for now is a group of input devices. Later this should be extended with a group of outputs and a group of buffer allocators/(cards)? For now this is a separate entity that shares ownership of the input devices with InputDeviceHub. In the upcoming steps the responsibility to send input device changes to the client is added.