Handle exceptions during platform construction
Since we no longer assign the mir::SharedLibrary for a graphics platform to a global variable or a DefaultServerConfiguration member, the SharedLibrary is destroyed when the unwinding leaves the_graphics_platform. When we rethrow the exception - to create detailed diagnostis - libstdc++ tries to access the unwinding information. But that has already been unmapped from memory.