Send the keymap instead of not sending the credentials over the wire
Moves the compilation of keymaps to the server, so that a client may never receive an invalid keymap set up. Also adds a MirDeviceId to the event, so that in a future iteration we can implement the per device keymaps. Those cannot be implemented right now, since the nested server has yet no access to the set of available devices that may need a keymap. 

In a better world we would use a mem-fd and compile the keymaps only once, and just send the keymap over the wire.