Add support for a wider range of mouse buttons, and ignore joystick buttons
Libinput currently only differs between keys and buttons. Buttons are always
treated as mouse buttons so we receive pointer button events in unexpected 
cases. With this change the pointer events with outside of the common mouse
range are ignored. Changes in libinput are required to get proper gamepad or
joystick support.