Define _GNU_SOURCE and _FILE_OFFSET_BITS=64 everywhere.
Various structures are defined differently when these are defined, meaning different parts of
the code saw different definitions.

When enabling LTO the linker is actually able to detect this, and throws ODR errors at us.

I don't believe that what we were doing was causing problems, but we *were* violating the ODR
rule, the errors are correct, and so we might as well fix them.