Make mt::DisplayConfigMatches composable.
The mt::DisplayConfigMatches matcher was using EXPECT_EQ under the covers, which is
OK as long as you only ever use it directly, but if you want to compose the matcher -
for example, by matching Not(mt::DisplayConfigMatches(foo)) - then it the EXPECT_EQ
calls cause the matcher to spuriously fail when the configurations (expectedly) don't match.

Likewise, this would break for any of the other matcher composers - AnyOf, and so on.