Disable key repeat on arale & mtk-tpd
The arale touchscreen driver simulates a button. When the screen is turned off while the button is held and then released mir will not receive the release event or any notice that the device is off. To avoid that mir keeps on sending repeat events, repeats for the touchscreen are disabled.