Improve readability of ClientLatency acceptance tests; source code and its results.
stats.post() has also been moved to avoid recording latencies of zero.
Our measurements remain coarse in that they are only frame numbers, but
this way it's clearer that a frame produced in the interval immediately
before it goes on screen has a latency of one and not zero. Especially
since render times should be closer to zero than vblank_interval.

I've also verified the test correctly detects if the 'early release'
feature is turned off. So after that happens we can tighten the threshold
again to not accept latencies above 2 (it's presently 3).