Server::run(): Destroy server config after only exception processing has completed.
The server config owns all UniqueModulePtr references to all the loaded platforms.
If the server config is destroyed, then all loaded platforms will be dlclos()ed.
Since we may be handling an exception thrown from one of those DSOs, this is unsafe
and can lead to a SIGEGV that appears to be stack corruption as the code it is
jumping to has been unloaded.