Store the cursor image configured
Due to rather accidental construction orders the initial configuration of the nested surface cursor images is lost.
It is done by mir::graphics::nested::Cursor::Cusor but in the current setup it is constructed ahead of mir::graphics::nested::Display and
any attached DisplayBuffers. With this change the construction order is no longer relevant.

A change of the construction order will happen due to the changes in the nested input platform.