Nested Input Platform
Convert the input setup code for a nested server into a platform. The platform is not dynamically loaded. Still part of the server, but only uses the HostConnection interface and the APIs use by other input platforms. As a result there is now an InputDeviceRegistry an input::Seat, thus the input setup is more similar to the host server.