kms-utils: Throw exceptions when DRM resources calls fail.
DRM resources are static; CRTCs, Connectors, and Encoders don't come and go.
All valid IDs are non-zero, so we can pre-check for validity rather than using
NULL to mean either “a call unexpectedly failed” or “the call failed because
it was asking for the invalid-guard”.

This makes debugging easier, as unexpected failures aren't hidden.