Fix GL_BGRA[_EXT] internal format confusion, again.
It appears more people believe this is the right combination, and the
Nvidia driver needs it this way to avoid software surfaces being all
black.

So this fixes Nvidia, but unfortunately does make windows all black for
software surfaces on desktop GL with Mesa. But apparently that's an open
Mesa bug: https://bugs.freedesktop.org/show_bug.cgi?id=92265

So THIS is the correct syntax according to the most sources, but
actually doesn't yet work (in desktop GL non-ES) for most people using
Mesa.