Avoid races in privileged test
Pull available devices out of udev on startup and not delayed when server is already started up.
In the privileged test it kind of depends on the speed of the device - i.e. the nested server could
connect to the host server before the evdev platform got around to look for available devices.