Remove superfluous unfocus-event-synthesis code.
It turns out that this code only ever gets executed by its unit test.

This is because (a) for the shell to process the destruction of a focused window it
must decide where the new focus goes. When the shell switches focus it pushes an
unfocus event onto the surface-being-destroyed's event sink. Then the Shell has finished
processing the destroy request we go back to SessionMediator and the response to the
destroy request is sent.

Because we process events in-order, the unfocus event will be processed before
the destroy-acknowledge message, so the surface cannot have the focused state
by the time we're destroying it client-side.