Fix server hang ups and allocation failures when client recreate buffer streams
When a mir client destroys a buffer stream the destruction result callback will trigger a buffer release of the associated buffers. That buffer release request inevitable fails in session mediator because it fails to resolve the BufferStreamId. This change handles exceptions during the attempt to disassociate buffers from already deleted buffer streams to ensure that the actual buffer release can happen.

This error is not really a leak - the buffers were deleted when the client disconnects and the server would be able to serve new clients.