android: make use of new PbufferGLContext constructor in DeviceQuirks
Originally, DeviceQuirks creates a PbufferGLContext with 
mir_pixel_format_abgr_8888 hard-coded as a format. This causes a problem
on a device without mir_pixel_format_abgr_8888 support. This revision
make use of new PbufferGLContext constructor which doesn't have to
specify a display format. This makes the code not depending on the
availability of mir_pixel_format_abgr_8888 format on the device.