Improvements to mtd::FakeDisplay.
Throw an exception when writing to/reading from the configuration change event trigger
fails rather than silently continuing, blithely unaware that the call has failed.

Switch from a pipe to an eventfd; eventfd is the solution tailored for this use-case.

Presented in the forlorn hope that it makes a test that appears to be flaky on vivid+clang,
but does not appear that it's possible for it to be flaky, stops being flaky.
Either by consistently passing, or consistently failing!