Fix pure virtual function call and segmentation fault
Lock access to display configuration in FakeDisplay and make sure that DisplayGroups are only accessed when compositing threads are allowed to schedule frames and not when the display mutex is not locked. Hence some initialization steps had to moved to be done synchronously by the initializing thread and other information used during startup and shutdown of the compositing thread is copied by value.