Improve the robustness of acceptance test:   TestClientInput.receives_one_touch_event_per_frame so that it is now less likely to fail on a heavily loaded or slow system. (LP: #1570698)
In reality we haven't seen this test actually fail for almost three months
but this is aimed at helping Brandon to get one of his branches to pass.

I actually started out trying to introduce fake server timestamps but found
they created a lot of failures under Valgrind, presumably because the client
wasn't using the same fake clock. But we should try to do that in future.