Make touch event creation a single step
This avoids another case in which we directly accessed the event type to update
a cookie. This is also a preparation step to make event validator not depend on
the internal structure of the input event types.

Additionally this removes touch contact size from the construction of touch
events as it was always pulled out of the major and minor axis sizes. Instead
this changes adds pieces to revive the contact orientation support in those
places that were touched by the change.