Make ObserverMultiplexer delegate observations to a mir::Executor.
This allows us to make ObserverMultiplexer vastly simpler, at the cost of moving
observer notifications out of the call stack.

The register_interest(Observer, Executor) overload will allow callers to get
observations via a mechanism of their choice - for example, on the GLib main-loop,
or on a Qt event loop.