TestPosixRWMutex: Resolve occasional deadlocks.
trigger_next_reader() would increment the current reader_to_run and then notify all waiters
that they might be ready to run. However, this could be run *twice* before the waiting thread was
scheduled. In that case, the waiting thread would wake up, notice that it's still not meant to be
running, and go back to sleep - we'd miss our turn to run.

So, make trigger_next_reader() idempotent.