deprecate unused and unsent event
InputConfigurationEvent was sent by the android EventHub and in cases we never hit. We never ported that behavior. For a similar purposes we do send a new bag of input devices via the protobuf connection, or for other similar purposes there are now input device state events - which do contain interesting information.