vanity: Add proper support for swap interval 0 (mirvanity -n)
Previously it would keep printing an error message that "your camera is
too slow" and also underestimate "expected range", but otherwise worked.

This change corrects all of that. The camera wasn't too slow, just that
the display is so fast with framedropping that the camera seems slow.
But it's still the same physical display so is only appropriate to get
the correct refresh rate from elsewhere. This also corrects the
"expected range" number.

I had to modify eglapp because I found there's no way for a user of
eglapp to get a SurfaceOutput event. The event has already arrived before
the app itself has initialized, so can only be acquired by the toolkit.