NestedServer tests: fix implicit dependecy on display configuration tearing down the compositor.
Some tests were waiting for nested display configuration to be applied before attaching any clients to the nested server.
Because the nested server doesn't composite anything until there's something to composite this meant that the nested
server did not have focus in the host server, and so its display configuration wouldn't get applied.

This was hidden because display configuration *previously* did a Compositor::stop()/Compositor::start() cycle,
which forces a composition pass, which would result in the nested server receiving focus.