Fix duplicated motion events on enter and filter out enter events in nested input platform
If the surface input dispatcher detects that the surface underneath the mouse cursor has changed it would send a leave event to the old window (containing the relative mouse motion) and the current position. The new window receives an enter event, and also the relative motion. After that the original event was sent. That led to sending the relative motion twice. It now depends on whether the original event contained a state change other than the motion itself. If so the motion change is removed from the original event, if not the event is discarded.