Fix races in SessionMediatorReportTest.
Reporting is no longer synchronous, so it's possible for the client to receive a reply before
the server report has been notified. The CI systems are sufficiently slow for this to
semi-regularly happen.

Wait for the reporting to happen rather than assume reporting has been done by the time the
client has recieved the response.