DisplayConfigurationTest: rely on fewer implementation details.
This changes apply_config_change_and_wait_for_propagation() to set the base configuration and then
wait for a configuration change notification advertising that it has been applied.

This has two advantages:
1) Setting the base configuration means that the configuration that gets set isn't filtered through
various policy bits which may modify some salient part of the configuration, and
2) The test is no longer tied to implementation details of MediatingDisplayChanger. Specifically,
it no longer assumes that MediatingDisplayChanger will synchronously push a configuration change
functor onto the server_action_queue() which will apply the configuration when run, and that the
server_action_queue runs in FIFO order.