Rename MIR_NDK_PATH to MIR_CHROOT.
The old name is now quite wrong and misleading. It hasn't been an NDK in
a long time. We just kept that name around for a while in case people
used the environment variable for their own caching, while automatic
caching was being introduced. Now automatic caching has been in place
for years (~/.cache/mir-${target_arch}-chroot-${dist}).

The new name is MIR_CHROOT rather than MIR_SYSROOT because it is only
set for chroot cross-compiling builds. If we want to we could add a
MIR_SYSROOT after this.