Enhance ClientLatency acceptance tests to support the possibility that the server may not render all client frames if it's under heavy load.
Instead, in the glorious future the server will just guarantee the
newest frame always reaches the screen. This allows it to catch up in
the face of a loaded system rather than visible lag increasing
indefinitely. It's also a feature of the coming client-side vsync that
the test needs to support in order to avoid spurious failures.